+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Usage:
 * Fix #2393: Remove timestamp from `org.label-schema.build-date` LABEL to utilize docker cache
 * Fix #2399: Helm no longer generates default function; broadens support for different value types
 * Fix #2400: Helm supports complex values in `values.yaml` fragments (such as annotations or arrays)
+* Fix #2414: OpenShift Gradle Plugin picks up `jkube.build.pushSecret` property
 
 _**Note**_:
 - Container Images generated using jkube opinionated defaults no longer contain full timestamp in `org.label-schema.build-date` label. The label contains the build date in the format `yyyy-MM-dd`.